requirements.txt
django
djangorestframework
drf-yasg
drf-spectacular
djangorestframework-simplejwt==5.2.2
